Update versions of other actions in README example
I noticed that these were out of date and I figure it couldn’t hurt to
keep the ancillary aspects of the example up to date, just in case
people copy-pasta the whole example.

Speaking of which, I was also tempted to add `cache: maven` to
the `setup-java` args since it seems like a good default to me, but
I held off just to keep this change focused on one specific kind of
change. That said, I do think that’d be a good idea.